# Use vector-tile C++ library