--- a/cmake/benchmark-files.cmake
+++ b/cmake/benchmark-files.cmake
@@ -9,6 +9,7 @@ set(MBGL_BENCHMARK_FILES
 
     # parse
     benchmark/parse/filter.benchmark.cpp
+    benchmark/parse/vector_tile.benchmark.cpp
 
     # src
     benchmark/src/main.cpp