--- a/cmake/benchmark.cmake
+++ b/cmake/benchmark.cmake
@@ -20,6 +20,7 @@ target_link_libraries(mbgl-benchmark
 target_add_mason_package(mbgl-benchmark PRIVATE benchmark)
 target_add_mason_package(mbgl-benchmark PRIVATE rapidjson)
 target_add_mason_package(mbgl-benchmark PRIVATE protozero)
+target_add_mason_package(mbgl-benchmark PRIVATE vector-tile)
 
 mbgl_platform_benchmark()