--- a/cmake/core.cmake
+++ b/cmake/core.cmake
@@ -27,6 +27,7 @@ target_add_mason_package(mbgl-core PRIVATE protozero)
 target_add_mason_package(mbgl-core PRIVATE polylabel)
 target_add_mason_package(mbgl-core PRIVATE wagyu)
 target_add_mason_package(mbgl-core PRIVATE shelf-pack)
+target_add_mason_package(mbgl-core PRIVATE vector-tile)
 
 mbgl_platform_core()