--- a/src/mbgl/tile/vector_tile_data.hpp
+++ b/src/mbgl/tile/vector_tile_data.hpp
@@ -41,14 +41,14 @@ public:
     VectorTileData(std::shared_ptr<const std::string> data);
 
     std::unique_ptr<GeometryTileData> clone() const override;
-    const GeometryTileLayer* getLayer(const std::string& name) const override;
+    std::unique_ptr<GeometryTileLayer> getLayer(const std::string& name) const override;
 
     std::vector<std::string> layerNames() const;
 
 private:
     std::shared_ptr<const std::string> data;
     mutable bool parsed = false;
-    mutable std::unordered_map<std::string, VectorTileLayer> layers;
+    mutable std::map<std::string, const protozero::data_view> layers;
 };
 
 } // namespace mbgl