… benchmark/parse/vector_tile.benchmark.cpp