This pull request replaces our custom vector tile parsing code with the [vector-tile C++ library](https://github.com/mapbox/vector-tile).

Addresses the problem in https://github.com/mapbox/mapbox-gl-native/issues/9167 by changing the `GeometryTileData` interface to return owning handles to `GeometryTileLayer`, which we can then retain in `SymbolLayout` so that the `SymbolFeature` handles don't end up with dangling handles. I opted not to make `GeometryTileFeature` handles fully owning their data due to possible overhead of a `shared_ptr`.

/cc @springmeyer