@kkaefer 2017-06-21
This will fail due to https://github.com/mapbox/vector-tile/pull/19

I added a benchmark and these are the results:

**before**

Benchmark                | Time         | CPU
---                      | ---          | ---
Parse_VectorTile         | 2.598.133 ns | 2.589.664 ns
Parse_VectorTile         | 2.562.716 ns | 2.558.520 ns
Parse_VectorTile         | 2.498.893 ns | 2.496.094 ns
Parse_VectorTile         | 2.645.567 ns | 2.637.008 ns
Parse_VectorTile         | 2.658.438 ns | 2.649.066 ns
Parse_VectorTile         | 2.628.378 ns | 2.620.168 ns
Parse_VectorTile         | 2.573.529 ns | 2.567.988 ns
Parse_VectorTile         | 2.570.820 ns | 2.567.393 ns
Parse_VectorTile         | 2.574.151 ns | 2.568.861 ns
Parse_VectorTile         | 2.560.927 ns | 2.555.828 ns
Parse_VectorTile_mean    | 2.587.155 ns | 2.581.059 ns
Parse_VectorTile_stddev  |    44.733 ns |    42.695 ns

**after**

Benchmark                | Time         | CPU
---                      | ---          | ---
Parse_VectorTile         | 1.922.978 ns | 1.908.669 ns
Parse_VectorTile         | 2.032.030 ns | 1.994.519 ns
Parse_VectorTile         | 1.943.548 ns | 1.920.645 ns
Parse_VectorTile         | 1.960.098 ns | 1.936.279 ns
Parse_VectorTile         | 1.992.238 ns | 1.973.358 ns
Parse_VectorTile         | 1.939.405 ns | 1.931.519 ns
Parse_VectorTile         | 1.919.359 ns | 1.916.416 ns
Parse_VectorTile         | 1.904.928 ns | 1.902.651 ns
Parse_VectorTile         | 1.930.251 ns | 1.927.384 ns
Parse_VectorTile         | 1.891.877 ns | 1.888.255 ns
Parse_VectorTile_mean    | 1.943.671 ns | 1.929.970 ns
Parse_VectorTile_stddev  |    39.796 ns |    30.510 ns