@springmeyer 2017-06-21
Great to see these perf improvements @kkaefer 🎉  Can you tell if the improvements are due to addressing #9167 or due to the upgraded vector tile decoding code? If it is the latter I would suspect/hope that it is attributable to either [how we now pre-allocate arrays for geometries](https://github.com/mapbox/vector-tile/commit/d131843974e54e5fe088b8ead78acd8bac12e615) or the new [deferred parsing of feature values until accessed (rather than at layer parse)](https://github.com/mapbox/vector-tile/commit/8faca0c2fab343cff42b071794c04395e1084fc1).