@kkaefer 2017-06-21
I ran Instruments, and it by far the dominating part is geometry parsing, and it looks like it's spending less time on that now.