@springmeyer 2017-06-21
> I ran Instruments, and it by far the dominating part is geometry parsing, and it looks like it's spending less time on that now.

👍 

@kkaefer I would recommend adding `getProperties` or `getValue` calls to the benchmark as well. The deferred `parseValue` will then get invoked there. And your benchmark would be more apples to apples to previously.