@kkaefer 2017-06-21
Added `getProperties()` to the test, and seeing a 3ms -> 2.3ms improvement.