@springmeyer 2017-06-21
> Added getProperties() to the test, and seeing a 3ms -> 2.3ms improvement.

👍  So, with `getProperties()` results were `~3ms` before and `~2.3` after the changes in this branch? Assuming I'm understanding, that is great news, and confirms we are going to be overall faster.